Split the grunt config blocks into grunt/ modules.
